Move everything arch-specific to arch submodules
